praefect/router: More eagerly verify selection of routing candidates
The tests for RouteRepositoryCreation try to verify that both the
primary and secondary nodes are picked from a certain pool of
candidates. This test is insufficient though as we only verify the
assumption when actually picking candidates. So in the case where we
don't even try to pick any, or when we try to pick multiple ones, then
we wouldn't notice and pass the test.

Fix this by instead allocating an array of available candidates that
gets appended to whenever a candidate is picked. This allows us to
easily assert the overall state after routing compared to only observing
snapshots and thus fixes the shortcomings documented above.