praefect/router: Move setup of repository store into testcases
The setup of the repository store for the RouteRepositoryCreation
function is hardwired into the test body. This makes it hard to add
tests that require a more involved setup.

Refactor the code to use an explicit setup function. While at it, move
hardcoded constants into the tests themselves so that it becomes easier
to reason about without having to switch back and forth between the
contants' declaration and the actual tests.