gitaly: Remove the `repositoryHook`
We use the `repositoryHook` to add a hook before the `ResolveRevision`
function is called. In reality the use-case is to add a hook before the
log is stored into the database.

In `1ff6dcd1` we introduced a framework for panics in the hooks. Let's
use this to have a `beforeStoreLogEntry` hook instead and remove the
`ResolveRevision` hook. This is more inline with what we actually want
to test.

Overall, this is done so we can remove the `repositoryHook` and remove
the `repository` interface altogether.