gitaly: Remove the `repository` interface
The interface `repository` was added to allow us to add hooks for
testing. In the previous commit, we removed this dependency. Let's
remove the interface altogether and pass `localrepo.Repo` everywhere. We
need this for the upcoming commits wherein we'll integrate
`housekeeping.Manager` which expects us to pass a `localrepo.Repo`
variable.