housekeeping: Default-enable geometric repacking
We have rolled out geometric repacking into production systems two weeks
ago now. Key observations:

    - On a global scale, we see a decrease in time spent optimizing
      repositories of 15-30%.

    - On a smaller scale, we see a decrease in time spent optimizing
      repositories in the gitlab-org and gitlab-com groups of 70-85%.

As geometric repacking is especially thought to be a win in large
monorepositories that have a lot of activity these observations match
the expectations that we have. Our gitlab-org and gitlab-com groups host
two of the most active repositories that are also quite large with the
gitlab and www-gitlab-com repositories, so it is expected that we see a
large win there.

On the global scale the win is expected to be less significant as we
tend to do full repacks more frequently in repositories with less
activity now. But given that we still see a win there this seems to not
be an issue.

Let's default-enable the feature flag so that we can remove it in the
next release.

Changelog: performance