repoutil: Extract RemoveRepository into repoutil
RemoveRepository is responsible for removing repositories in a
transactional race-free way. To implement backups on the server, instead
of calling the RemoveRepository RPC, it would be preferrable to have
function we could call. So here we extract the implementation of
RemoveRepository into repoutil. This package already contains the
implementation for CreateRepository.