git/stats: Convert alternates into a structure
The `RepositoryInfo` structure has a field that exposes alternate object
directories to the caller. In the context of repository housekeeping
we'll also need the information when the alternates file was last
modified.

Let's prepare for this change by converting the simple field into a
structure `AlternatesInfo` that is self-contained. Add tests that
exercise parsing of alternates more thoroughly, where some of the tests
highlight buggy behaviour.