Makefile: Add target for building proto docs
The old `gitaly-proto` published a GitLab Pages site with documentation
for the Gitaly protobufs. This is useful for users outside of the Gitaly
team to better understand RPCs without digging into the repo.

To start migrating this functionality over to the Gitaly repo, add a
target to the Makefile to build HTML documentation for our protos.

Changelog: added