Remove gitaly_pack_objects_limiting_remote_ip feature flag
This feature flag controls the rollout of pack-objects concurrency limit
by remote IP. We rolled this feature out on GitLab.com production and
received positive results. This commit removes this feature flag.

Afterward, the concurrency limit for pack-objects command invocation is
always enabled. The default configuration is 200 per IP per node. The
loopback IP is excluded from the limiter. Requests exceeding this limit
are forced to be queued. They can stay in the queue as long as they
want.