grpc-proxy: Restructure and move package into `internal/grpc/`
We've grown quite a bunch of packages that provide various different
functionality in the context of gRPC. But it is not easy to identify
them as there is no common hierarchy for gRPC-related functionality.
We're thus introducing a new `internal/grpc` package that will host
various bits and pieces related to gRPC.

Move the `internal/praefect/grpc-proxy` package into this new package.

As we have originally vendored this project from upstream, this package
does not quite match our typical layout: it's got scripts and docs in
there that don't really make much sense nowadays. Originally, we had the
intent to contribute all changes back upstream. But the project is
basically dead anyway, and by now we have significantly diverged and
added Praefect-specific logic that wouldn't be upstreamable.

So let's get rid of this cruft while at it. While it is debatable
whether things like the readme or examples should go away I very much
doubt that anybody ever looks at these files. Instead, we should really
provide good documentation on the public interfaces and provide package
level documentation like we already do.