Do not use a separate error variable for mergeErr
It is more idiomatic to overwrite the customary `err` variable with
each new error encountered in a function, instead of
having a separate variable for each potential error.
This function already followed this pattern, with a single
exception: `mergeErr`. The exception had already caused a bug,
because an error message made inside the `mergeErr`
handler was using `err` out of habit.