Wire PartitionManager into operations.Server in tests
PartitionManager is the entry point for the WAL logic. It handles
routing transaction to the correct partitions and the lifecycle of
TransactionManagers responsible for those partitions. As a first step
into integrating it into our tests, wire the PartitionManager into
OperationService tests. As we eventually want run the tests with both
WAL enabled and disabled, the PartitionManager is configured only if
'GITALY_TEST_WAL' environment variable is set. All of the users of the
WAL will have to check whether or not it is set. This remains the same
when we eventually make it possible to enable the WAL logic through the
configuration. If set, we'll configure and dependency inject. If not set,
we'll leave it nil.

testPackObjectsConcurrency was changed to use a unique config for each
test case so each of the tests have a unique storages for the PartitionManager
to initialize state in.