gittest: Introduce wrapper to retrive object-hash dependent values
Although we should and do try to avoid having object-hash dependent
test logic, it can't be avoided at times. Most of the callsites that
have hash dependent logic fail open though, where they fall back to
either the SHA1 or SHA256 value in case a completely unknown object
format is in use.

Introduce a new helper function that handles this more explicitly. The
function accepts a map of values keyed by the object format so that it
is readily extensible if the need ever arises. While this isn't really
all that likely, it is still nice to explicitly annotate all callsites
that would need to be amended.