housekeeping: Convert `findStaleReferenceLocks` to a closure
The function `findStaleReferenceLocks` is used to find and clear stale
reference locks in a given repository against a grace period. This grace
period is hardcoded and clients of `CleanStaleData` cannot configure it.

To provide this configuration option, we convert
`findStaleReferenceLocks` to a closure which takes in a grace period as
an argument.