housekeeping: Introduce and use `CleanStaleDataConfig`
We introduce a new struct `CleanStaleDataConfig`, which is used to hold
the configuration around which cleanup functions are to be run inside
`CleanStaleData`.

With this we also introduce a function `DefaultStaleDataCleanup` which
will provide the default configuration as it currently exists. We then
use this in `CleanStaleDataConfig` to provide the configuration.

In following commits we will make this an argument to
`CleanStaleDataConfig`.