housekeeping: Make `CleanStaleData` configurable
In previous commits we introduced a configuration for `CleanStaleData`
via the struct `CleanStaleDataConfig`. In this commit, we make this
configuration an argument to `CleanStaleData`. This allow users to
supply their own configuration and only run the cleanup functions they
require.

We fix tests around this and also add a test to test a non-default grace
period for `findStaleReferenceLocks`.