localrepo: Extract FetchBundle implementation from FetchBundle RPC
Extracts the implementation of the FetchBundle RPC so that it can be
used for server-side backups.

Having to pass in transaction manager is probably an abstraction leak
since localrepo is about wrapping git, not gitaly functionality. I think
we should eventually create a "gitaly repo" abstraction that handles
transactions when this cannot be done through git hooks. There are
already some of these functions clustered in repoutil.

CreateRepositoryFromBundle was using some common functionality, and so
this implementation got absorbed at the same time.