Break storages in tests only after initialization
We're about to wire PartitionManager into Gitaly's test server. On
initialization, it creates expected directories and opens a database
handle for each storage's database. This does not work in some tests
as they configure invalid storages which don't exist. Let's change the
tests to break the storages after initialization. Gitaly should anyway
be ensure that the storages exist on boot.