client: Allow overriding transport credentials
The `Dial()` function automatically sets up transport credentials based
on the schema of the passed address. For insecure connections it will
set up insecure credentials, while for secure connections it will load
the system certificate pool. The consequence is that even when callers
pass in their own transport credentials via the gRPC dial options, they
will get overridden.

While seemingly okay in the context of production systems, it does cause
issues in the context of testing: we often want to generate certificates
at test time and use them, but we're forced to somehow inject them into
the system certificate pool. This requires fiddling with environment
variables and is overall not great.

Introduce a new option that allows the caller to override the transport
credentials.