Split UpdaterWithHooks into a separate package
We're about to integrate Gitaly's WAL logic into the first tests.
This is about to create a cyclic import between the ìnternal/gitaly
and internal/git/updateref packages. TransactionManager uses the
updateref package to perform the refrence updates. UpdaterWithHooks
is used from the operations service package to commit reference
changes and execute hooks. In order to handle Transaction's from the
TransactionManager, the updateref package would have to import them
creating a cyclic dependency.

UpdaterWithHooks is only used from Gitaly's services, and is not really
a utility for working with Git itself. Move it's implementation to
internal/gitaly/hook/updateref to break the cycle. It's places as a
subpackage in the hook package as it uses it to run the hooks.