praefect: Refactor server factory test to be less repetitive
One of our Praefect server factory tests sets up multiple different
sockets, checks each of them for health and then asserts that stopping
the server closes the sockets. This logic is quite repetitive and thus
easy to get wrong, as the preceding commit demonstrates.

Refactor the test to instead set up the sockets in a loop.