Wire Transaction to the HookManager
Transaction has a read snapshot that includes a specific version of
custom hooks. TransactionManager is managing hooks via multiversion
concurrency control, that is, hooks are always written as an entirely
new version and never update in place. The HookManager needs to take
in the Transaction so it can determine which version of the custom
hooks to execute. This commit wires the Transaction to the HookManager
but doesn't do any logic changes yet.