Extract hook path generation into a function
This commit extracts the hook path generation into a function.
Extracting a function will later make it easier in tests to assert
we have the correct path when we add the hook path to the transaction's
snapshot.

While at it, update the parent directory to be synced with the
recently introduced helper designed for it.