Include custom hook path in a transaction's snapshot
TransactionManager manages custom hooks via MVCC. Custom hooks are always
written into a new directory to isolate transactions from concurrent writes.
When executing or reading the custom hooks, the transactions should use
the custom hooks included in the snapshot. While the log index of the
custom hooks is already included in the snapshot, it's not useful for
integration as it doesn't yet say where exactly the custom hooks should be
executed from. Solve this problem by including an absolute path to the
custom hooks on the disk. This will later be used by HookManager to execute
the correct custom hooks for a transaction, and GetCustomHooks to fetch the
correct version of the custom hooks.

Backwards compatibility logic is included to execute custom hooks from the
repository if none have yet been written via WAL.