Include hook path in a transaction's snapshot
TransactionManager manages hooks via MVCC. Hooks are always written
into a new directory to isolate transactions from concurrent writes.
When executing or reading the hooks, the transactions should use
the hooks included in the snapshot. While the log index of the hooks
is already included in the snapshot, it's not useful for integration
as it doesn't yet say where exactly the hooks should be executed from.
Solve this problem by including an absolute path to the hooks on the
disk. This will later be used by HookManager to execute the correct
hooks for a transaction.

Backwards compatibility logic is included to execute hooks from the
repository if none have yet been written via WAL.