Execute custom hooks from transaction's snapshot
Each transaction has a read snapshot that maintains its isolation
from concurrent writes. The snapshot includes also the version of
custom hooks. When the hooks are executed or read, the operation
should be targeted at the correct version of hooks. The transaction's
snapshot includes the absolute path to the custom hooks on the disk.
This commit teaches HookManager to execute the snapshotted hooks
from the directory included in the transaction.

While the HookManager knows how to execute the correct hooks now,
not all of the call sites pass a transaction yet.