praefect: Fix flaky server factory test
The server factory test that verifies whether we correctly shut down
sockets when stopping the server is flaky:

    === FAIL: internal/praefect TestServerFactory/stops_all_listening_servers (0.11s)
        server_factory_test.go:281:
                    Error Trace:	/builds/gitlab-org/gitaly/internal/praefect/server_factory_test.go:281
                                                            /usr/local/go/src/runtime/asm_amd64.s:1594
                    Error:      	Received unexpected error:
                                    grpc: the server has been stopped
                    Test:       	TestServerFactory/stops_all_listening_servers
        --- FAIL: TestServerFactory/stops_all_listening_servers (0.11s)

The error here indicates that the server has been stopped before we
started to listen on it. Theoretically speaking this shouldn't happen
though as we connect to the server and query its health service, so it
must have been ready to serve. But there is a bug in this test: we spawn
multiple sockets and serve on all of them, but accidentally connect to
the TCP listener twice instead of hitting the TCP and TLS listeners.

Fix this bug by connecting to the correct socket address. This also
surfaces that we haven't been setting up transport credentials as
required.