tools/test-boot: Gracefully terminate Gitaly
When Gitaly has booted up we kill it by sending it a SIGKILL. This will
immediately reap the process without any cleanup logic running, which
has the consequence that the process will not clean up after itself.

Fix this by sending a SIGTERM instead. This is a bit slower, but not
really so much that it would matter:

    Benchmark 1: go run ./tools/test-boot
      Time (mean ± σ):      3.033 s ±  0.143 s    [User: 3.623 s, System: 0.263 s]
      Range (min … max):    2.954 s …  3.409 s    10 runs

    Benchmark 2: go run ./tools/test-boot
      Time (mean ± σ):      3.113 s ±  0.074 s    [User: 3.645 s, System: 0.292 s]
      Range (min … max):    2.979 s …  3.224 s    10 runs

    Summary
      'go run ./tools/test-boot' ran
        1.03 ± 0.05 times faster than 'go run ./tools/test-boot'

A 3% performance hit should certainly be acceptable.