tools/test-boot: Bump timeout to 30 seconds
The test-boot utility waits up to 10 seconds for the Gitaly server to
start accepting connections. We have seen in our CI systems that this is
seemingly not sufficient anymore as there are frequent failures in our
build step where we hit this timeout. And that isn't quite surprising:
test-booting Gitaly on a beefy machine already takes around 3 seconds,
so it's expected that a slow and busy CI machine might potentially take
a lot longer than that.

Let's bump the timeout to 30 seconds to avoid these flaky CI runs.