Allow forcing reference updates in TransactionManager
TransactionManager is currently always verifying the reference's old
value matches the expected one in the update. There are some RPCs in
Gitaly that perform updates forcefully without checking the reference's
old value. One such example is WriteRef. To support such use cases, this
commit implements a Force flag in ReferenceUpdate that allows for applying
the reference update regardless of the old value.