go.mod: Update urfave/cli module to v2.25.3
While the Renovate bot has created the update to urfave/cli v2.25.3 for
us already, it was closed due to a slight change in the output it
generates when required flags are missing: instead of printing the
single-letter abbreviation, the module now prints the complete flag
name. This was upstream changed in a7e3459 (fix: missing required flag
error uses flag name and not alias, 2023-03-20).

Update the urfave/cli module to v2.25.3 and fix the now-broken test.