cli/praefect: Fix flaky tests for the metadata subcommand
The tests for the "metadata" subcommand fail reliably on my machine,
seemingly due to a race condition:

```
    subcmd_metadata_test.go:134:
                Error Trace:	.../internal/cli/praefect/subcmd_metadata_test.go:134
                Error:      	Not equal:
                                expected: <nil>(<nil>)
                                actual  : *fmt.wrapError(&fmt.wrapError{msg:"get metadata: rpc error: code = Internal desc = get metadata: query: driver: bad connection", err:(*status.Error)(0xc0009044e8)})
                Test:       	TestMetadataSubcommand/repository_found_with_repository_id
    --- FAIL: TestMetadataSubcommand/repository_found_with_repository_id (0.00s)
```

The root cause seems to the that the transaction that we're using gets
into a bad state once a query run inside of it returns an error, which
in fact happens in one of the subtests because the queried repository
does not exist.

Fix this flake by creating per-subtest transactions. Note that we can't
just stop using transactions because health checks may otherwise start
to become flaky.