Export HeadReference from localrepo
localrepo currently exports GetDefaultBranch for getting the
repository's default branch. This is generally the value of HEAD.
GetDefaultBranch also has some legacy logic and returns a different
value in some cases. In cases where one really needs to have the
actual value of HEAD, let's export the HeadReference method to do so.
This will eventually be used in the TransactionManager to determine
the value of repository's HEAD.