Implement reference update voting in TransactionManager
Praefect's transactions are currently relying on Git's
reference-transaction hook to vote on the changes. The hook
verifies the references and locks them before voting. When WAL
is enabled, this needs to be handled in the TransactionManager
instead as Git's reference transaction hook is not part of the
write flow anymore. This commit implements voting for reference
updates in TransactionManager.

The semantics stay the same with the references verified prior to
voting and concurrent operations prevented. There is no explicit
locking though as the synchronization relies on the TransactionManager
processing only a single transaction at a time. This leads to coarser
locking. Where as there could previously be multiple transactions
voting concurrently on disjoint reference updates, with the current
implementation in TransactionManager there can only be one transaction
voting at a time. This leads to worse performance but should behaviorally
be correct. We can later improve the performance with more fine grained
locking semantics. The focus right now is having a functional
implementation that can be integrated and tested.

The voting logic performs the vote on the whole transaction which could
produce votes that the previous voting logic wouldn't, say if references
are updated in the same transaction as the default branch. In practice
this doesn't matter, as none of the RPC handlers would commit such a
transaction.

The voting logic is not tested in separately in the TransactionManager's
tests. Most importantly we should test that the votes produced by the
RPCs are the same with the WAL logic and non-WAL logic. The votes
themselves are just opaque hashes, so we can't really say much about
whether they are correct or not in isolation. We'll thus test the logic
only through integration in the RPC handlers.