Add a helper for logging rollback errors
Rolling back a transaction may return an error. This commit adds a
helper function that rolls back the transaction and logs any possible
error returned. This makes integrating the WAL more convenient as one
can simply defer this after beginning a transaction to roll it back
in case of errors and log the error.